simplify onContractPicked method
